+++ frontend/package.json
@@ -13,7 +13,7 @@
         "tsc:noEmit": "tsc --noEmit",
         "clean": "rm -rf coverage build",
         "test": "concurrently --kill-others-on-fail npm:jest npm:eslint npm:prettier npm:tsc -c green,blue,cyan,magenta",
-        "jest": "JEST_JUNIT_OUTPUT_DIR=coverage jest test --runInBand --all --watchAll=false --ci --coverage --reporters=default --coverageReporters=lcov --coverageReporters=html --coverageReporters=text --coverageReporters=text-summary --testResultsProcessor jest-sonar-reporter",
+        "jest": "JEST_JUNIT_OUTPUT_DIR=coverage jest --runInBand --all --watchAll=false --ci",
         "jest:parallel": "jest",
         "eslint": "eslint src --ext .js,.jsx,.ts,.tsx --max-warnings=0",
         "eslint:fix": "eslint src --ext .js,.jsx,.ts,.tsx --fix",
@@ -37,7 +37,7 @@
         "@kubernetes/client-node": "0.16.3",
         "@octokit/rest": "18.12.0",
         "@octokit/types": "6.34.0",
-        "@patternfly-labs/react-form-wizard": "^1.7.5",
+        "@patternfly-labs/react-form-wizard": "^1.7.8",
         "@patternfly/react-code-editor": "4.49.6",
         "@patternfly/react-core": "4.208.6",
         "@patternfly/react-icons": "4.59.6",
